Update Discovery method
